Update README.md
Its a python program that lets user to block access to any url from system for either safety or personal reason. It is beneficial for child safety, limited access to internet, and others. User can scheduled it on startup of system by task scheduler program. Program runs according to given time frame. That is in given time frame program will block access to blacklisted website and out of that it allow them.